Tasklist toolbar
Added a toolbar to the tasklist screen, allowing for deleting selected tasks, updating selectedtasks with a status and adding a new task. Styling still needs to be applied.

I will try and get the datagrid search options in the left side of the toolbar, but this may not be possible.